requirements.txt
APScheduler==3.9.1
autopep8==1.7.0
backports.zoneinfo==0.2.1
boto3==1.26.5
botocore==1.29.5
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.3
cssbeautifier==1.14.6
djlint==1.17.3
EditorConfig==0.12.3
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.14.6
MarkupSafe==2.1.1
pathspec==0.10.1
pycodestyle==2.9.1
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.41
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
ec2_metadata==2.10.0
